requirements2.txt
boto3==1.12.31
botocore==1.15.31
certifi==2019.9.11
chardet==3.0.4
cloudinary==1.20.0
cycler==0.10.0
defusedxml==0.6.0
Django==2.2.6
django-allauth==0.40.0
django-cloudinary-storage==0.2.3
django-environ==0.4.5
django-storages==1.9.1
docutils==0.15.2
idna==2.8
jmespath==0.9.5
kiwisolver==1.1.0
matplotlib==3.1.1
numpy==1.17.3
oauthlib==3.1.0
opencv-python==4.1.1.26
Pillow==6.2.1
psycopg2==2.8.3
psycopg2-binary==2.8.3
pyparsing==2.4.2
python-dateutil==2.8.0
python3-openid==3.1.0
pytz==2019.3
requests==2.22.0
requests-oauthlib==1.2.0
s3transfer==0.3.3
six==1.12.0
sqlparse==0.3.0
urllib3==1.25.6
